requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
mysql-connector-python==8.0.27
protobuf==3.19.1
pydantic==1.8.2
python-dotenv==0.19.2
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
